+++ src/paco/models/base.py
@@ -155,18 +155,12 @@ class Parent(CFNExport):
         "List of paco.ref parts"
         obj = self
         parts = []
-
-        # special case for global buckets
-        if get_parent_by_interface(obj, schemas.IGlobalResources) and schemas.IS3Bucket.providedBy(obj):
-            account = obj.account.split('.')[-1:][0]
-            parts = ['resource','s3','buckets', account, obj.region, obj.name]
-        else:
+        parent = obj.__parent__
+        while parent != None:
+            parts.append(obj.name)
+            obj = parent
             parent = obj.__parent__
-            while parent != None:
-                parts.append(obj.name)
-                obj = parent
-                parent = obj.__parent__
-            parts.reverse()
+        parts.reverse()
         return parts
 
     @property